requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
passlib==1.7.4
pathlib==1.0.1
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pymongo==4.7.3
requests==2.31.0
uvicorn==0.29.0
